requirements.txt
django
djangorestframework
django-oauth-toolkit
gunicorn
django-heroku